refactor: Move logic for popup matcher to separate class
Test reviews

Fix children matching
