+++ kakao/src/main/kotlin/com/agoda/kakao/spinner/SpinnerAdapterActions.kt
@@ -34,12 +34,10 @@ interface SpinnerAdapterActions {
                     size = view.count
                 }
             }
-
         })
 
         return size
     }
-
 
     /**
      * Opens the spinner to display it's items
