+++ sample/src/androidTest/kotlin/com/agoda/sample/SpinnerTest.kt
@@ -3,6 +3,7 @@ package com.agoda.sample
 import androidx.test.internal.runner.junit4.AndroidJUnit4ClassRunner
 import androidx.test.rule.ActivityTestRule
 import com.agoda.kakao.screen.Screen.Companion.onScreen
+import com.agoda.kakao.spinner.KSpinnerItem
 import com.agoda.sample.screen.SpinnerActivityScreen
 import org.junit.Rule
 import org.junit.Test
@@ -23,15 +24,13 @@ class SpinnerTest {
 
                 open()
 
-                emptyFirstChild {
+                childAt<KSpinnerItem>(0) {
                     isVisible()
-
-                    hasText("Title 1")
+                    hasText("Title 0")
                 }
 
                 emptyLastChild {
                     isVisible()
-
                     hasText("Title 10")
                 }
 
@@ -39,10 +38,6 @@ class SpinnerTest {
                     isInstanceOf(String::class.java)
                     equals("Title 5")
                 }
-//                perform {
-//                    isDisplayed()
-//                    hasText("Title 5")
-//                }
 
                 emptyChildAt(4) {
                     isDisplayed()
