+++ sample/src/main/kotlin/com/agoda/sample/SpinnerActivity.kt
@@ -11,13 +11,7 @@ class SpinnerActivity : AppCompatActivity() {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_spinner)
 
-        val adapter = ArrayAdapter<String>(
-            this, android.R.layout.simple_spinner_item,
-            listOf(
-                "Title 1", "Title 2", "Title 3", "Title 4", "Title 5",
-                "Title 6", "Title 7", "Title 8", "Title 9", "Title 10"
-            )
-        )
+        val adapter = ArrayAdapter(this, android.R.layout.simple_spinner_item, (0..9).map { "Title $it" })
         adapter.setDropDownViewResource(android.R.layout.simple_spinner_dropdown_item)
         val spinner = findViewById<Spinner>(R.id.spinner)
         spinner.adapter = adapter
